// src/api/enums/auth/auth-error-code.enum.ts
export enum AuthErrorCode {
  EMAIL_NOT_VERIFIED = 'EMAIL_NOT_VERIFIED',
  INVALID_VERIFICATION_CODE = 'INVALID_VERIFICATION_CODE',
  VERIFICATION_TOKEN_INVALID = 'VERIFICATION_TOKEN_INVALID',
  VERIFICATION_TOKEN_EXPIRED = 'VERIFICATION_TOKEN_EXPIRED',
  RESEND_TOO_SOON = 'RESEND_TOO_SOON',
  TOO_MANY_ATTEMPTS = 'TOO_MANY_ATTEMPTS',
  INVALID_RESET_CODE = 'INVALID_RESET_CODE',
  RESET_CODE_NOT_VERIFIED = 'RESET_CODE_NOT_VERIFIED',
  RESET_TOKEN_INVALID = 'RESET_TOKEN_INVALID',
  RESET_TOKEN_EXPIRED = 'RESET_TOKEN_EXPIRED',
}

export const AuthErrorCodeEnumName = 'AuthErrorCode';